# backend/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book Management System API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/authors": {
            "post": {
                "description": "生年月日は現在より過去日付",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "著者を作成する",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateAuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "put": {
                "description": "生年月日は現在より過去日付。version が一致しない場合は 409",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "著者を更新する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "著者ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateAuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/authors/{id}/books": {
            "get": {
                "description": "書籍が見つからない場合は空配列を返す",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "著者に紐づく書籍一覧を取得する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "著者ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AuthorBookResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/books": {
            "post": {
                "description": "書籍価格は0以上、著者は1人以上（重複不可・存在する著者のみ）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "書籍を作成する",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "put": {
                "description": "出版済みから未出版には変更できない。version が一致しない場合は 409",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "書籍を更新する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "書籍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "AuthorBookResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "integer",
                    "example": 1500
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "人間失格"
                }
            }
        },
        "AuthorResponse": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "type": "string",
                    "example": "1909-06-19"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "太宰治"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "BookAuthorResponse": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "type": "string",
                    "example": "1909-06-19"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "太宰治"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "BookResponse": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/BookAuthorResponse"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "integer",
                    "example": 1500
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "人間失格"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "CreateAuthorRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "birthDate": {
                    "type": "string",
                    "example": "1909-06-19"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "太宰治"
                }
            }
        },
        "CreateBookRequest": {
            "type": "object",
            "required": [
                "authorIds",
                "price",
                "status",
                "title"
            ],
            "properties": {
                "authorIds": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 0,
                    "example": 1500
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "人間失格"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Field-level details; only set for request validation errors.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FieldError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "invalid"
                }
            }
        },
        "FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "title"
                },
                "rule": {
                    "type": "string",
                    "example": "required"
                }
            }
        },
        "UpdateAuthorRequest": {
            "type": "object",
            "required": [
                "name",
                "version"
            ],
            "properties": {
                "birthDate": {
                    "type": "string",
                    "example": "1909-06-19"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "太宰治"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "UpdateBookRequest": {
            "type": "object",
            "required": [
                "authorIds",
                "price",
                "status",
                "title",
                "version"
            ],
            "properties": {
                "authorIds": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 0,
                    "example": 1500
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "人間失格"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}